% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatial_correlation.R
\name{check_spatial_correlation}
\alias{check_spatial_correlation}
\title{Mantel correlograms}
\usage{
check_spatial_correlation(p, var_names = NULL, n.class = 10, nperm = 100)
}
\arguments{
\item{p}{\code{SpatVector} object}

\item{var_names}{\code{character} vector containing the names of the variables in \code{p}.}

\item{n.class}{Number of classes (see \code{vegan}).}

\item{nperm}{Number of permutations for the tests of significance (see \code{vegan}).}
}
\value{
A \code{mantel.correlog} object.
}
\description{
\code{check_spatial_correlation} calculates Mantel correlograms
}
