% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{datos_especie}
\alias{datos_especie}
\title{Title}
\usage{
datos_especie(carpeta, nombre_especie, ifn = 2)
}
\arguments{
\item{carpeta}{Cadena de caracteres indicando la carpeta en la que encontrar los archivos
de datos ".rds" de los IFN.}

\item{nombre_especie}{Cadena de caracteres con el nombre en latín de la especie.}

\item{ifn}{Valor numérico indicando qué IFN hay que leer. Debe ser 2, 3 o 4.}
}
\value{
Un objecto "sf" con el contenido de los datos de presencia de la especie.
}
\description{
Title
}
\examples{
carpeta <- "C:/imidra/ifn"
nombre_especie <- "Cistus ladanifer"
df <- datos_especie(carpeta, nombre_especie)
}
