% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_occurrence.R
\name{extract_occurrence}
\alias{extract_occurrence}
\title{EXtract single occurrence file}
\usage{
extract_occurrence(path, species, target = "tree", verbose = TRUE)
}
\arguments{
\item{path}{\code{character} string indicating the path to the file with the data..}

\item{species}{\code{character} string with the full name of the species.}

\item{target}{\code{character} string specifying what to extract. It can be
"tree", "shrub", "herbs" or "regen".}

\item{verbose}{\code{logical} variable, if set to TRUE (default) progress information
is shown on screen.}
}
\value{
description
A \code{data.frame} containing a logical column labelled "occurrence".
All other columns are the same that are found in the original ".rds" file.
}
\description{
\code{extract_occurrence} extracts occurrence data from a single IFN data file
and for a single species. That occurrence data can be for a tree, shrub, herb and
regenerate species.
}
\details{
Straightforward.
}
\examples{
# IFN data.
path <- "C:/imidra/peninsula/ifn/ifn4_28.rds"
species <- "Erica arborea"
x <- extract_occurrence(path, species, "shrub")

}
