% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_occurrence_list.R
\name{extract_occurrence_list}
\alias{extract_occurrence_list}
\title{Title}
\usage{
extract_occurrence_list(
  path,
  species,
  target = "tree",
  merge = TRUE,
  crs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{\code{character} string indicating the path to the file with the data..}

\item{species}{\code{character} string with the full name of the species.}

\item{target}{\code{character} string specifying what to extract. It can be
"tree", "shrub", "herbs" or "regen".}

\item{merge}{\code{logical} if set to TRUE}

\item{crs}{\code{character} string specifying a valid EPSG code.}

\item{verbose}{\code{logical} variable, if set to TRUE (default) progress information
is shown on screen.}
}
\value{
An object, or list of objects, containing a logical column labelled "occurrence".
All other columns are the same that are found in the original ".rds" file.

There are three options for the output:
\enumerate{
\item merge = FALSE and crs = NULL: a list of data.frames as obtained from 'readRDS'
\item merge = FALSE and crs = valid crs: a list of SpatVect objects projected to the input 'crs'.
\item merge = TRUE and crs = valid crs; a single SpatVect object containing all files and projected to 'crs'.
}
}
\description{
\code{extract_occurrence_list} extracts occurrence data from several IFN data files
and for a single species. That occurrence data can be for a tree, shrub, herb and
regenerate species.
}
\examples{
# IFN data.
path <- file.path("C:/imidra/peninsula/ifn", paste0("ifn4_", formatC(1:50, width = 2, flag = "0"), ".rds"))
species <- "Erica arborea"
path <- select_files_exist(path)
x <- extract_occurrence_list(path, species, "shrub", merge = TRUE, crs = "EPSG:4326")
}
