% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_predictors.R
\name{extract_predictors}
\alias{extract_predictors}
\title{Title}
\usage{
extract_predictors(p, x, verbose = TRUE)
}
\description{
Title
}
\examples{
folder <- "C:/imidra"
species <- "Cistus ladanifer"
p <- get_presence(folder, species)

# Absolute paths to files.
carpetas <- list(mdt = file.path("C:/imidra", "mdt/mdt_madrid.tif"),
bioclim = file.path("C:/imidra/bioclim", paste0("historico/wc2.1_10m_bio/wc2.1_10m_bio_", 1:19, ".tif")),
corine = file.path("C:/imidra", "corine/corine_2018/CORINE Madrid nivel 1.tif"),
hidro = file.path("C:/imidra", "hidro/distancia_hidro.tif"))

# Name the bioclimatic variables.
names(carpetas$bioclim) <- paste0("bioclim_", 1:19)

# Read data.
x <- get_predictors(carpetas)

# Extract predictors at 'p' locations.
y <- extract_predictors(p, x)
}
