% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioclim_spain_future.R
\name{get_bioclim_spain_future}
\alias{get_bioclim_spain_future}
\title{Title}
\usage{
get_bioclim_spain_future(
  base_url_to_download = NULL,
  gcm = "EC-Earth3-Veg",
  ssp = 126,
  resolution = "30s",
  destination_folder = NULL,
  epsg = NULL,
  mainland = TRUE,
  verbose = TRUE
)
}
\description{
Title
}
