% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon_spain.R
\name{get_polygon_spain}
\alias{get_polygon_spain}
\title{Title}
\usage{
get_polygon_spain(epsg = NULL, mainland = TRUE)
}
\description{
Title
}
\details{
Simple use of package 'mapSpain' to retrieve polygons.
}
