% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{Predictor variables at presence locations.}
\usage{
get_predictors(folder, species, verbose = TRUE)
}
\arguments{
\item{folder}{\code{character} string with the root name of the folder with the data.}

\item{species}{\code{character} string with the full name of the species.}

\item{verbose}{\code{logical}, if TRUE progress information is produced.}
}
\value{
A \code{terra} SpatVector object.
}
\description{
\code{get_predictors} extracts the value of the different predictor variables
that are used in the analysis at all presence locations for the input
species.
}
\examples{
folder <- "C:/imidra/"
species <- "Cistus ladanifer"
x <- get_predictors(folder, species)
}
