% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_variables}
\alias{select_variables}
\title{Variable selection with hierarchical clustering}
\usage{
select_variables(
  x,
  stability = TRUE,
  B = 50,
  k = NULL,
  cutoff = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{data.frame} or \code{sf} object containing a set of variables.}

\item{stability}{\code{logical} variable, if set to TRUE stability of partition is calculated.}

\item{B}{\code{numeric}, number of bootstrap samples.}

\item{verbose}{\code{logical} if set to TRUE, information about progress is printed.}
}
\value{
A \code{list} containing the final model, a \code{data.frame} with the percentage of homogeneity
that is accounted by each number of clusters and the subset of data variables. If \code{stability}
is set to TRUE the output of the stability analysis is added.
}
\description{
\code{select_variables} uses hierarchical clustering from package \code{ClustOfVar}
to select best subset of variables.
}
