% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_predictors.R
\name{selecting_predictors}
\alias{selecting_predictors}
\title{Title}
\usage{
selecting_predictors()
}
\description{
Title
}
\examples{
# Read predictor data.
# Extract predictors at 'p' locations.
}
