% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_filter.R
\name{spatial_filter}
\alias{spatial_filter}
\title{Title}
\usage{
spatial_filter(p, min_dist = 20, num_simu = 10)
}
\arguments{
\item{min_dist}{\code{numeric} value, minimum distance in Km.}
}
\description{
Title
}
