% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{Predictor variables at presence locations.}
\usage{
get_predictors(carpeta, especie)
}
\arguments{
\item{carpeta}{\code{character} string}

\item{especie}{}
}
\value{
A \code{terra} SpatVector object.
}
\description{
\code{get_predictors} extracts the value of the different predictor variables
that are used in the analysis at all presence locations for the input
species.
}
\examples{
carpeta <- "C:/imidra/"
nombre_especie <- "Cistus ladanifer"
df <- get_predictors(carpeta, nombre_especie)
}
