% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get_predictors}
\alias{get_predictors}
\title{Predictor variables at presence locations.}
\usage{
get_predictors(path, crs = NULL, verbose = TRUE)
}
\arguments{
\item{crs}{\code{character} string containing the EPSG code (in the form "EPSG:####) of the
coordinate reference system all raster should be projected to. If not given, or set to NULL,
no projection will be carried out.
#' @param verbose \code{logical}, if TRUE progress information is produced.}
}
\value{
A \code{terra} SpatRasterCollection object.
}
\description{
\code{get_predictors} extracts the value of the different predictor variables
that are used in the analysis at all presence locations for the input
species.
}
\examples{
# Absolute paths to files.

# Name the bioclimatic variables.
names(carpetas$bioclim) <- paste0("bioclim_", 1:19)

# Read data.
x <- get_predictors(carpetas)
}
