% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence}
\alias{get_presence}
\title{Title}
\usage{
get_presence(folder, species, ifn = 4, verbose = TRUE)
}
\arguments{
\item{folder}{\code{character} string indicating the root folder to retrieve the data from.}

\item{species}{\code{character} string with the full name of the species.}

\item{ifn}{\code{numeric} value indicating which IFN to read. It can be equal to 2, 3 or 4.}
}
\value{
An "sf" object containing presence data for the species.
}
\description{
Title
}
\examples{
folder <- "C:/imidra"
species <- "Cistus ladanifer"
species <- "Erica arborea"
x <- get_presence(folder, species)

}
