% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence}
\alias{get_presence}
\title{Title}
\usage{
get_presence(path, species, crs = NULL, verbose = TRUE)
}
\arguments{
\item{path}{\code{character} string indicating the path to the file with the data..}

\item{species}{\code{character} string with the full name of the species.}
}
\value{
An "SpatVector" object containing presence data for the species.
}
\description{
Title
}
\examples{
# Madrid data.
path <- "C:/imidra/ifn/ifn4_28.rds"
species <- "Cistus ladanifer"
species <- "Erica arborea"
x <- get_presence(path, species, "EPSG:25830")

}
