% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_epsg.R
\name{is_valid_epsg}
\alias{is_valid_epsg}
\title{Title}
\usage{
is_valid_epsg(code)
}
\arguments{
\item{code}{\code{character} string of the type 'epsg:####', where #### is a valid EPSG numeric code.}
}
\value{
TRUE or FALSE.
}
\description{
Title
}
\examples{
is_valid_epsg("epsg:3")
is_valid_epsg("epsg:4286")

# Error.
is_valid_epsg("3")
}
