% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_null_simu.R
\name{maxent_null_simu}
\alias{maxent_null_simu}
\title{Title}
\usage{
maxent_null_simu(
  mask,
  raster_list,
  num_presence,
  background,
  min_dist = 1,
  num_null = 100,
  size = 1000,
  verbose = TRUE
)
}
\arguments{
\item{np}{\code{numeric} number of presence points.}

\item{bx}{spatially filtered background SpatVector object.}
}
\description{
Title
}
