% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_variables}
\alias{select_variables}
\title{Variable selection with hierarchical clustering}
\usage{
select_variables(
  x,
  stability = TRUE,
  B = 50,
  k = NULL,
  cutoff = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{data.frame} or \code{sf} object containing a set of variables.}

\item{stability}{\code{logical} variable, if set to TRUE stability of partition is calculated.}

\item{B}{\code{numeric}, number of bootstrap samples.}

\item{verbose}{\code{logical} if set to TRUE, information about progress is printed.}
}
\value{
A \code{list} containing the final model, a \code{data.frame} with the percentage of homogeneity
that is accounted by each number of clusters and the subset of data variables. If \code{stability}
is set to TRUE the output of the stability analysis is added.
}
\description{
\code{select_variables} uses hierarchical clustering from package \code{ClustOfVar}
to select best subset of variables.
}
\examples{
# All numeric data.
data(decathlon, package = "ClustOfVar")
subset <- select_variables(decathlon, cutoff = 0.95)
plot(subset$homogeneity, xlab = "N. of clusters", ylab = "Gain in homogeneity")
points(c(0, 13), c(.95, .95), type = "l", lty = 2, lwd = 2)

# All categorical data.
data(vnf, package = "ClustOfVar")
subset <- select_variables(vnf, cutoff = 0.95)
plot(subset$homogeneity, xlab = "N. of clusters", ylab = "Gain in homogeneity")
points(c(0, 13), c(.95, .95), type = "l", lty = 2, lwd = 2)


folder <- "C:/imidra"
species <- "Cistus ladanifer"
p <- get_presence(folder, species)

# Absolute paths to files.
carpetas <- list(mdt = file.path("C:/imidra", "mdt/mdt_madrid.tif"),
bioclim = file.path("C:/imidra/bioclim", paste0("historico/wc2.1_10m_bio/wc2.1_10m_bio_", 1:19, ".tif")),
corine = file.path("C:/imidra", "corine/corine_2018/CORINE Madrid nivel 1.tif"),
hidro = file.path("C:/imidra", "hidro/distancia_hidro.tif"))

# Name the bioclimatic variables.
names(carpetas$bioclim) <- paste0("bioclim_", 1:19)

# Read data.
x <- get_predictors(carpetas)

# Extract predictors at 'p' locations.
p <- extract_predictors(p, x)

# Variable selection.
subset <- select_variables(p, cutoff = 0.95)
}
